Isolate video overlay from engine facade
overlay_video is a bounded picture-in-picture transform and can move into a focused module while preserving mcp_video.engine.overlay_video for public callers. The extraction keeps the existing behavior but sanitizes scale, opacity, and timing filter values before FFmpeg interpolation.

Constraint: Split overlay from split_screen and composite-overlay work to avoid overlapping the remaining high-risk composition paths.

Rejected: Extract split_screen in the same PR | separate filter graphs and test surfaces make it safer as its own slice.

Confidence: high

Scope-risk: moderate

Directive: Keep FFmpeg filter values sanitized before interpolation; position handling relies on _resolve_position validation.

Tested: ruff check mcp_video/engine.py mcp_video/engine_overlay.py; ruff format --check mcp_video/engine.py mcp_video/engine_overlay.py; python3 -m pytest tests/test_engine_advanced.py -k 'overlay' -q --tb=short; python3 -m pytest tests/test_cli.py -k 'overlay_video' tests/test_server.py -k 'overlay' tests/test_red_team.py -k 'overlay' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
